backend/antigravity/brain_service/requirements.txt
pika
psycopg2-binary
pgvector
redis
pydantic
supabase
